requirements.txt
annotated-types==0.7.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.0
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
